@@ -42,58 +42,59 @@
     "update": "ng update @nrwl/workspace",
     "workspace-schematic": "nx workspace-schematic",
     "dep-graph": "nx dep-graph",
-    "help": "nx help"
+    "help": "nx help",
+    "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points"
   },
   "dependencies": {
-    "@angular/animations": "^8.2.0",
-    "@angular/common": "^8.2.0",
-    "@angular/compiler": "^8.2.0",
-    "@angular/core": "^8.2.0",
-    "@angular/forms": "^8.2.0",
-    "@angular/platform-browser": "^8.2.0",
-    "@angular/platform-browser-dynamic": "^8.2.0",
-    "@angular/router": "^8.2.0",
-    "@fortawesome/angular-fontawesome": "^0.5.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.25",
-    "@fortawesome/free-brands-svg-icons": "^5.11.2",
-    "@fortawesome/free-solid-svg-icons": "^5.11.2",
-    "@nrwl/angular": "8.12.2",
+    "@angular/animations": "9.1.2",
+    "@angular/common": "9.1.2",
+    "@angular/compiler": "9.1.2",
+    "@angular/core": "9.1.2",
+    "@angular/forms": "9.1.2",
+    "@angular/platform-browser": "9.1.2",
+    "@angular/platform-browser-dynamic": "9.1.2",
+    "@angular/router": "9.1.2",
+    "@fortawesome/angular-fontawesome": "^0.6.1",
+    "@fortawesome/fontawesome-svg-core": "^1.2.28",
+    "@fortawesome/free-brands-svg-icons": "^5.13.0",
+    "@fortawesome/free-solid-svg-icons": "^5.13.0",
+    "@nrwl/angular": "9.2.3",
     "core-js": "^2.5.4",
-    "ngx-bootstrap": "^5.3.2",
+    "ngx-bootstrap": "^5.5.0",
     "ngx-markdown": "^9.0.0",
-    "rxjs": "~6.4.0",
-    "zone.js": "^0.9.1"
+    "rxjs": "6.5.5",
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.23",
-    "@angular-devkit/build-ng-packagr": "~0.803.3",
-    "@angular/cli": "8.3.14",
-    "@angular/compiler-cli": "^8.2.0",
-    "@angular/language-service": "^8.2.0",
+    "@angular-devkit/build-angular": "0.901.2",
+    "@angular-devkit/build-ng-packagr": "0.901.2",
+    "@angular/cli": "9.1.2",
+    "@angular/compiler-cli": "9.1.2",
+    "@angular/language-service": "9.1.2",
     "@commitlint/cli": "^8.3.5",
     "@commitlint/prompt": "^8.3.5",
-    "@nrwl/cypress": "8.12.2",
-    "@nrwl/workspace": "8.12.2",
-    "@types/jest": "24.0.9",
+    "@nrwl/cypress": "9.2.3",
+    "@nrwl/workspace": "9.2.3",
+    "@types/jest": "25.1.4",
     "@types/node": "^8.9.5",
-    "codelyzer": "~5.0.1",
-    "commitizen": "^4.0.3",
+    "codelyzer": "~5.2.2",
+    "commitizen": "^4.0.4",
     "commitlint": "^8.3.5",
     "conventional-changelog-cli": "^2.0.31",
     "cypress": "^3.8.2",
     "dotenv": "6.2.0",
-    "eslint": "6.1.0",
+    "eslint": "6.8.0",
     "husky": "^4.2.3",
-    "jest": "24.1.0",
-    "jest-preset-angular": "7.0.0",
-    "ng-packagr": "^5.4.0",
-    "prettier": "1.18.2",
-    "ts-jest": "24.0.0",
+    "jest": "25.2.4",
+    "jest-preset-angular": "8.1.3",
+    "ng-packagr": "^9.1.1",
+    "prettier": "1.19.1",
+    "ts-jest": "25.2.1",
     "ts-node": "~7.0.0",
-    "tsickle": "~0.35.0",
-    "tslint": "~5.11.0",
+    "tsickle": "~0.38.1",
+    "tslint": "6.0.0",
     "typedoc": "^0.16.9",
-    "typescript": "~3.5.3"
+    "typescript": "3.8.3"
   },
   "config": {
     "commitizen": {
