@@ -61,4 +61,4 @@ import 'zone.js/dist/zone'; // Included with Angular CLI.
  * APPLICATION IMPORTS
  */
 
-import 'core-js/es7/reflect';
+import 'core-js/es6/reflect';
